Refactor product serialization helpers for API and service
